removing code duplicates and replace unsigned long by uint32_t

+++ modules/features2d/src/agast_score.hpp
@@ -53,7 +53,7 @@ namespace cv
 {
 
 #if !(defined __i386__ || defined(_M_IX86) || defined __x86_64__ || defined(_M_X64))
-int agast_tree_search(const unsigned long table_struct32[], int pixel_[], const unsigned char* const ptr, int threshold);
+int agast_tree_search(const uint32_t table_struct32[], int pixel_[], const unsigned char* const ptr, int threshold);
 int AGAST_ALL_SCORE(const uchar* ptr, const int pixel[], int threshold, int agasttype);
 #endif //!(defined __i386__ || defined(_M_IX86) || defined __x86_64__ || defined(_M_X64))
